refactor(zipkin) make extractor.lua into a handler.lua aux function
